+++ include/Classifier.h
@@ -42,27 +42,27 @@ class Classifier {
         stream >> m_numberOfFlatnessFeatures;
         stream >> m_can_use_fast_forest;
         m_fast_forest = readForestFromStream<float>(stream);
-        m_binned_forest = readForestFromStream<unsigned int>(stream);
+        m_binned_forest = readForestFromStream<unsigned long>(stream);
 
       }
 
       friend std::ostream& operator<<(std::ostream& stream, const Classifier& classifier);
 
-			Classifier(unsigned int nTrees, unsigned int depth, std::vector<unsigned int> binning, double shrinkage = 0.1, double subsample = 1.0, bool sPlot = false, double flatnessLoss = -1.0, std::vector<bool> purityTransformation = {}, unsigned int numberOfFlatnessFeatures=0, bool transform2probability=true) :
+			Classifier(unsigned long nTrees, unsigned long depth, std::vector<unsigned long> binning, double shrinkage = 0.1, double subsample = 1.0, bool sPlot = false, double flatnessLoss = -1.0, std::vector<bool> purityTransformation = {}, unsigned long numberOfFlatnessFeatures=0, bool transform2probability=true) :
         m_nTrees(nTrees), m_depth(depth), m_binning(binning), m_shrinkage(shrinkage), m_subsample(subsample), m_sPlot(sPlot), m_flatnessLoss(flatnessLoss), m_purityTransformation(purityTransformation), m_numberOfFlatnessFeatures(numberOfFlatnessFeatures), m_transform2probability(transform2probability), m_can_use_fast_forest(true) { }
 
       void Print();
 
-      unsigned int GetNTrees() const { return m_nTrees; }
-      void SetNTrees(unsigned int nTrees) { m_nTrees = nTrees; }
+      unsigned long GetNTrees() const { return m_nTrees; }
+      void SetNTrees(unsigned long nTrees) { m_nTrees = nTrees; }
       
-      unsigned int GetDepth() const { return m_depth; }
-      void SetDepth(unsigned int depth) { m_depth = depth; }
+      unsigned long GetDepth() const { return m_depth; }
+      void SetDepth(unsigned long depth) { m_depth = depth; }
       
-      unsigned int GetNumberOfFlatnessFeatures() const { return m_numberOfFlatnessFeatures; }
-      void SetNumberOfFlatnessFeatures(unsigned int numberOfFlatnessFeatures) { m_numberOfFlatnessFeatures = numberOfFlatnessFeatures; }
+      unsigned long GetNumberOfFlatnessFeatures() const { return m_numberOfFlatnessFeatures; }
+      void SetNumberOfFlatnessFeatures(unsigned long numberOfFlatnessFeatures) { m_numberOfFlatnessFeatures = numberOfFlatnessFeatures; }
 
-      unsigned int GetNFeatures() const { return m_numberOfFeatures; }
+      unsigned long GetNFeatures() const { return m_numberOfFeatures; }
 
       double GetShrinkage() const { return m_shrinkage; }
       void SetShrinkage(double shrinkage) { m_shrinkage = shrinkage; }
@@ -76,8 +76,8 @@ class Classifier {
       bool GetTransform2Probability() const { return m_transform2probability; }
       void SetTransform2Probability(bool transform2probability) { m_transform2probability = transform2probability; }
       
-      std::vector<unsigned int> GetBinning() const { return m_binning; }
-      void SetBinning(std::vector<unsigned int> binning) { m_binning = binning; }
+      std::vector<unsigned long> GetBinning() const { return m_binning; }
+      void SetBinning(std::vector<unsigned long> binning) { m_binning = binning; }
 
       std::vector<bool> GetPurityTransformation() const { return m_purityTransformation; }
       void SetPurityTransformation(std::vector<bool> purityTransformation) { m_purityTransformation = purityTransformation; }
@@ -89,34 +89,34 @@ class Classifier {
 
       float predict(const std::vector<float> &X) const;
       
-      std::map<unsigned int, double> GetVariableRanking() const;
+      std::map<unsigned long, double> GetVariableRanking() const;
       
-      std::map<unsigned int, double> GetIndividualVariableRanking(const std::vector<float> &X) const;
+      std::map<unsigned long, double> GetIndividualVariableRanking(const std::vector<float> &X) const;
 
-      std::map<unsigned int, unsigned int> GetFeatureMapping() const;
+      std::map<unsigned long, unsigned long> GetFeatureMapping() const;
   
-      std::map<unsigned int, double> MapRankingToOriginalFeatures(std::map<unsigned int, double> ranking) const;
+      std::map<unsigned long, double> MapRankingToOriginalFeatures(std::map<unsigned long, double> ranking) const;
 
   private:
-    unsigned int m_version = 1;
-    unsigned int m_nTrees = 100;
-    unsigned int m_depth = 3;
-    std::vector<unsigned int> m_binning;
+    unsigned long m_version = 1;
+    unsigned long m_nTrees = 100;
+    unsigned long m_depth = 3;
+    std::vector<unsigned long> m_binning;
     double m_shrinkage = 0.1;
     double m_subsample = 0.5;
     bool m_sPlot = true;
     double m_flatnessLoss = -1;
     std::vector<bool> m_purityTransformation;
-    unsigned int m_numberOfFlatnessFeatures = 0;
+    unsigned long m_numberOfFlatnessFeatures = 0;
     bool m_transform2probability = true;
-    unsigned int m_numberOfFeatures = 0;
-    unsigned int m_numberOfFinalFeatures = 0;
+    unsigned long m_numberOfFeatures = 0;
+    unsigned long m_numberOfFinalFeatures = 0;
     std::vector<FeatureBinning<float>> m_featureBinning;
     std::vector<PurityTransformation> m_purityBinning;
 
     bool m_can_use_fast_forest = true;
     Forest<float> m_fast_forest;
-    Forest<unsigned int> m_binned_forest;
+    Forest<unsigned long> m_binned_forest;
 
 };
